requirements.txt
discord.py
aiohttp
python-dotenv
pyyaml
gspread
oauth2client
asyncpg